feat: make money exact at the persistence and accounting domain layers
- switch `MONEY`/`SHARES` to `asdecimal=True` and add a `RATE` numeric type for the six former `double precision` rate columns
- add `db.money` as the single scale, rounding, and quantize policy, plus `RateMap` for the JSONB allocation map
- add `accounting.ledger.frame` as the named T1 boundary where exact money becomes the float analytics projection
- move the ledger frame schema off `Posting`, separating the exact domain model from the float projection
- make importer and paystub parsing exact from the source cell onward
- replace the split-validation and paystub tolerances with exact comparison
- pin the wire representation to JSON `number`, leaving the OpenAPI contract byte-identical; PR3 owns the decimal-string cutover

+++ src/accounting/db/budgets.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import uuid
+from decimal import Decimal
 from typing import get_args
 
 from sqlalchemy import CheckConstraint, ForeignKey, Index, UniqueConstraint, text
@@ -50,7 +51,7 @@ class Budget(Base):
     subcategory_id: Mapped[uuid.UUID | None] = mapped_column(
         UUID(as_uuid=True), ForeignKey(f"{SCHEMA}.categories.id"), default=None
     )
-    amount: Mapped[float] = mapped_column(MONEY)
+    amount: Mapped[Decimal] = mapped_column(MONEY)
     currency: Mapped[str] = mapped_column(default="USD")
 
 
@@ -85,5 +86,5 @@ class GeneralBudget(Base):
     subcategory_id: Mapped[uuid.UUID | None] = mapped_column(
         UUID(as_uuid=True), ForeignKey(f"{SCHEMA}.categories.id"), default=None
     )
-    amount: Mapped[float] = mapped_column(MONEY)
+    amount: Mapped[Decimal] = mapped_column(MONEY)
     currency: Mapped[str] = mapped_column(default="USD")
